// taskflow-graphql/src/generated/graphql.ts
export type Attachment = {
  __typename?: 'Attachment';
  data: Scalars['String']['output'];
  id: Scalars['ID']['output'];
  name: Scalars['String']['output'];
  size: Scalars['Int']['output'];
  storagePath?: Maybe<Scalars['String']['output']>;
  type: Scalars['String']['output'];
  uploadedAt: Scalars['DateTime']['output'];
};

export type AttachmentInput = {
  data: Scalars['String']['input'];
  name: Scalars['String']['input'];
  size: Scalars['Int']['input'];
  type: Scalars['String']['input'];
};

export type Board = {
  __typename?: 'Board';
  columns: Array<BoardColumn>;
  completedTaskCount: Scalars['Int']['output'];
  createdAt: Scalars['DateTime']['output'];
  description?: Maybe<Scalars['String']['output']>;
  id: Scalars['ID']['output'];
  isShared: Scalars['Boolean']['output'];
  name: Scalars['String']['output'];
  settings: BoardSettings;
  taskCount: Scalars['Int']['output'];
  updatedAt: Scalars['DateTime']['output'];
};

export type BoardColumn = {
  __typename?: 'BoardColumn';
  color?: Maybe<Scalars['String']['output']>;
  id: Scalars['ID']['output'];
  position: Scalars['Int']['output'];
  taskCount: Scalars['Int']['output'];
  title: Scalars['String']['output'];
};

export type BoardColumnInput = {
  color?: InputMaybe<Scalars['String']['input']>;
  id?: InputMaybe<Scalars['ID']['input']>;
  position: Scalars['Int']['input'];
  title: Scalars['String']['input'];
};

export type BoardSettings = {
  __typename?: 'BoardSettings';
  autoArchiveCompleted: Scalars['Boolean']['output'];
  completedColumnId?: Maybe<Scalars['ID']['output']>;
  defaultColumnId?: Maybe<Scalars['ID']['output']>;
  recycleBinRetentionDays: Scalars['Int']['output'];
};
// ...
export type Label = {
  __typename?: 'Label';
  boardId?: Maybe<Scalars['ID']['output']>;
  color: Scalars['String']['output'];
  createdAt: Scalars['DateTime']['output'];
  id: Scalars['ID']['output'];
  name: Scalars['String']['output'];
  taskCount: Scalars['Int']['output'];
};

export enum MarkdownFormat {
  GithubFlavored = 'GITHUB_FLAVORED',
  Obsidian = 'OBSIDIAN',
  Standard = 'STANDARD',
}

export type MarkdownMetadata = {
  __typename?: 'MarkdownMetadata';
  boardName: Scalars['String']['output'];
  completedCount: Scalars['Int']['output'];
  includeAttachments: Scalars['Boolean']['output'];
  includeLabels: Scalars['Boolean']['output'];
  includeSubtasks: Scalars['Boolean']['output'];
  taskCount: Scalars['Int']['output'];
};

export type MarkdownReport = {
  __typename?: 'MarkdownReport';
  content: Scalars['String']['output'];
  format: MarkdownFormat;
  generatedAt: Scalars['DateTime']['output'];
  metadata: MarkdownMetadata;
};

export type MarkdownReportInput = {
  boardId: Scalars['ID']['input'];
  filters?: InputMaybe<TaskFilters>;
  format?: InputMaybe<MarkdownFormat>;
  includeAttachments?: InputMaybe<Scalars['Boolean']['input']>;
  includeCompleted?: InputMaybe<Scalars['Boolean']['input']>;
  includeLabels?: InputMaybe<Scalars['Boolean']['input']>;
  includeSubtasks?: InputMaybe<Scalars['Boolean']['input']>;
};

export type Mutation = {
  __typename?: 'Mutation';
  breakdownTask: Array<Task>;
  createBoard: Board;
  createLabel: Label;
  createTask: Task;
  createTaskFromNaturalLanguage: Task;
  createTasks: Array<Task>;
  createTemplate: Template;
  createWebhook: Webhook;
  deleteBoard: Scalars['Boolean']['output'];
  deleteLabel: Scalars['Boolean']['output'];
  deleteTask: Scalars['Boolean']['output'];
  deleteTasks: Scalars['Boolean']['output'];
  deleteTemplate: Scalars['Boolean']['output'];
  deleteWebhook: Scalars['Boolean']['output'];
  duplicateTask: Task;
  generateMarkdownReport: MarkdownReport;
  moveTask: Task;
  optimizeTaskSchedule: ScheduleOptimization;
  restoreTask: Task;
  testWebhook: WebhookDelivery;
  updateBoard: Board;
  updateLabel: Label;
  updateTask: Task;
  updateTasks: Array<Task>;
  updateTemplate: Template;
  updateWebhook: Webhook;
};
// ...
export type PriorityBreakdown = {
  __typename?: 'PriorityBreakdown';
  critical: Scalars['Int']['output'];
  high: Scalars['Int']['output'];
  low: Scalars['Int']['output'];
  medium: Scalars['Int']['output'];
};

export type Query = {
  __typename?: 'Query';
  aiSuggestedTasks: Array<Task>;
  board?: Maybe<Board>;
  boards: Array<Board>;
  currentBoard?: Maybe<Board>;
  exportBoardAsMarkdown: Scalars['String']['output'];
  exportTaskAsMarkdown: Scalars['String']['output'];
  exportTasksAsMarkdown: Scalars['String']['output'];
  label?: Maybe<Label>;
  labels: Array<Label>;
  nextRecommendedTask?: Maybe<Task>;
  searchTasksByNaturalLanguage: Array<Task>;
  task?: Maybe<Task>;
  taskStatistics: TaskStatistics;
  tasks: Array<Task>;
  template?: Maybe<Template>;
  templates: Array<Template>;
  webhook?: Maybe<Webhook>;
  webhookDeliveries: Array<WebhookDelivery>;
  webhookDelivery?: Maybe<WebhookDelivery>;
  webhookStats: WebhookStats;
  webhooks: Array<Webhook>;
};
// ...
export type RecurrenceConfig = {
  __typename?: 'RecurrenceConfig';
  dayOfMonth?: Maybe<Scalars['Int']['output']>;
  dayOfWeekInMonth?: Maybe<Scalars['Int']['output']>;
  daysOfWeek?: Maybe<Array<Scalars['Int']['output']>>;
  enabled: Scalars['Boolean']['output'];
  endDate?: Maybe<Scalars['DateTime']['output']>;
  interval: Scalars['Int']['output'];
  maxOccurrences?: Maybe<Scalars['Int']['output']>;
  pattern: RecurrencePattern;
  weekOfMonth?: Maybe<Scalars['Int']['output']>;
};
// ...
export type AiSuggestionResolvers<
  ContextType = GraphQLContext,
  ParentType extends
    ResolversParentTypes['AISuggestion'] = ResolversParentTypes['AISuggestion'],
> = {
  actions?: Resolver<
    Array<ResolversTypes['SuggestedAction']>,
    ParentType,
    ContextType
  >;
  confidence?: Resolver<ResolversTypes['Float'], ParentType, ContextType>;
  message?: Resolver<ResolversTypes['String'], ParentType, ContextType>;
  task?: Resolver<Maybe<ResolversTypes['Task']>, ParentType, ContextType>;
  type?: Resolver<ResolversTypes['SuggestionType'], ParentType, ContextType>;
  __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
};

export type AttachmentResolvers<
  ContextType = GraphQLContext,
  ParentType extends
    ResolversParentTypes['Attachment'] = ResolversParentTypes['Attachment'],
> = {
  data?: Resolver<ResolversTypes['String'], ParentType, ContextType>;
  id?: Resolver<ResolversTypes['ID'], ParentType, ContextType>;
  name?: Resolver<ResolversTypes['String'], ParentType, ContextType>;
  size?: Resolver<ResolversTypes['Int'], ParentType, ContextType>;
  storagePath?: Resolver<
    Maybe<ResolversTypes['String']>,
    ParentType,
    ContextType
  >;
  type?: Resolver<ResolversTypes['String'], ParentType, ContextType>;
  uploadedAt?: Resolver<ResolversTypes['DateTime'], ParentType, ContextType>;
  __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
};

export type BoardResolvers<
  ContextType = GraphQLContext,
  ParentType extends
    ResolversParentTypes['Board'] = ResolversParentTypes['Board'],
> = {
  columns?: Resolver<
    Array<ResolversTypes['BoardColumn']>,
    ParentType,
    ContextType
  >;
  completedTaskCount?: Resolver<ResolversTypes['Int'], ParentType, ContextType>;
  createdAt?: Resolver<ResolversTypes['DateTime'], ParentType, ContextType>;
  description?: Resolver<
    Maybe<ResolversTypes['String']>,
    ParentType,
    ContextType
  >;
  id?: Resolver<ResolversTypes['ID'], ParentType, ContextType>;
  isShared?: Resolver<ResolversTypes['Boolean'], ParentType, ContextType>;
  name?: Resolver<ResolversTypes['String'], ParentType, ContextType>;
  settings?: Resolver<ResolversTypes['BoardSettings'], ParentType, ContextType>;
  taskCount?: Resolver<ResolversTypes['Int'], ParentType, ContextType>;
  updatedAt?: Resolver<ResolversTypes['DateTime'], ParentType, ContextType>;
  __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
};

export type BoardColumnResolvers<
  ContextType = GraphQLContext,
  ParentType extends
    ResolversParentTypes['BoardColumn'] = ResolversParentTypes['BoardColumn'],
> = {
  color?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>;
  id?: Resolver<ResolversTypes['ID'], ParentType, ContextType>;
  position?: Resolver<ResolversTypes['Int'], ParentType, ContextType>;
  taskCount?: Resolver<ResolversTypes['Int'], ParentType, ContextType>;
  title?: Resolver<ResolversTypes['String'], ParentType, ContextType>;
  __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
};

export type BoardSettingsResolvers<
  ContextType = GraphQLContext,
  ParentType extends
    ResolversParentTypes['BoardSettings'] = ResolversParentTypes['BoardSettings'],
> = {
  autoArchiveCompleted?: Resolver<
    ResolversTypes['Boolean'],
    ParentType,
    ContextType
  >;
  completedColumnId?: Resolver<
    Maybe<ResolversTypes['ID']>,
    ParentType,
    ContextType
  >;
  defaultColumnId?: Resolver<
    Maybe<ResolversTypes['ID']>,
    ParentType,
    ContextType
  >;
  recycleBinRetentionDays?: Resolver<
    ResolversTypes['Int'],
    ParentType,
    ContextType
  >;
  __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
};

export interface DateTimeScalarConfig
  extends GraphQLScalarTypeConfig<ResolversTypes['DateTime'], any> {
  name: 'DateTime';
}

export interface JsonScalarConfig
  extends GraphQLScalarTypeConfig<ResolversTypes['JSON'], any> {
  name: 'JSON';
}

export type LabelResolvers<
  ContextType = GraphQLContext,
  ParentType extends
    ResolversParentTypes['Label'] = ResolversParentTypes['Label'],
> = {
  boardId?: Resolver<Maybe<ResolversTypes['ID']>, ParentType, ContextType>;
  color?: Resolver<ResolversTypes['String'], ParentType, ContextType>;
  createdAt?: Resolver<ResolversTypes['DateTime'], ParentType, ContextType>;
  id?: Resolver<ResolversTypes['ID'], ParentType, ContextType>;
  name?: Resolver<ResolversTypes['String'], ParentType, ContextType>;
  taskCount?: Resolver<ResolversTypes['Int'], ParentType, ContextType>;
  __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
};

export type MarkdownMetadataResolvers<
  ContextType = GraphQLContext,
  ParentType extends
    ResolversParentTypes['MarkdownMetadata'] = ResolversParentTypes['MarkdownMetadata'],
> = {
  boardName?: Resolver<ResolversTypes['String'], ParentType, ContextType>;
  completedCount?: Resolver<ResolversTypes['Int'], ParentType, ContextType>;
  includeAttachments?: Resolver<
    ResolversTypes['Boolean'],
    ParentType,
    ContextType
  >;
  includeLabels?: Resolver<ResolversTypes['Boolean'], ParentType, ContextType>;
  includeSubtasks?: Resolver<
    ResolversTypes['Boolean'],
    ParentType,
    ContextType
  >;
  taskCount?: Resolver<ResolversTypes['Int'], ParentType, ContextType>;
  __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
};

export type MarkdownReportResolvers<
  ContextType = GraphQLContext,
  ParentType extends
    ResolversParentTypes['MarkdownReport'] = ResolversParentTypes['MarkdownReport'],
> = {
  content?: Resolver<ResolversTypes['String'], ParentType, ContextType>;
  format?: Resolver<ResolversTypes['MarkdownFormat'], ParentType, ContextType>;
  generatedAt?: Resolver<ResolversTypes['DateTime'], ParentType, ContextType>;
  metadata?: Resolver<
    ResolversTypes['MarkdownMetadata'],
    ParentType,
    ContextType
  >;
  __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
};
// ...
export type PriorityBreakdownResolvers<
  ContextType = GraphQLContext,
  ParentType extends
    ResolversParentTypes['PriorityBreakdown'] = ResolversParentTypes['PriorityBreakdown'],
> = {
  critical?: Resolver<ResolversTypes['Int'], ParentType, ContextType>;
  high?: Resolver<ResolversTypes['Int'], ParentType, ContextType>;
  low?: Resolver<ResolversTypes['Int'], ParentType, ContextType>;
  medium?: Resolver<ResolversTypes['Int'], ParentType, ContextType>;
  __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
};
// ...
export type Resolvers<ContextType = GraphQLContext> = {
  AISuggestion?: AiSuggestionResolvers<ContextType>;
  Attachment?: AttachmentResolvers<ContextType>;
  Board?: BoardResolvers<ContextType>;
  BoardColumn?: BoardColumnResolvers<ContextType>;
  BoardSettings?: BoardSettingsResolvers<ContextType>;
  DateTime?: GraphQLScalarType;
  JSON?: GraphQLScalarType;
  Label?: LabelResolvers<ContextType>;
  MarkdownMetadata?: MarkdownMetadataResolvers<ContextType>;
  MarkdownReport?: MarkdownReportResolvers<ContextType>;
  Mutation?: MutationResolvers<ContextType>;
  PriorityBreakdown?: PriorityBreakdownResolvers<ContextType>;
  Query?: QueryResolvers<ContextType>;
  RecurrenceConfig?: RecurrenceConfigResolvers<ContextType>;
  ScheduleOptimization?: ScheduleOptimizationResolvers<ContextType>;
  StatusBreakdown?: StatusBreakdownResolvers<ContextType>;
  SubTask?: SubTaskResolvers<ContextType>;
  Subscription?: SubscriptionResolvers<ContextType>;
  SuggestedAction?: SuggestedActionResolvers<ContextType>;
  Task?: TaskResolvers<ContextType>;
  TaskStatistics?: TaskStatisticsResolvers<ContextType>;
  TaskTemplateData?: TaskTemplateDataResolvers<ContextType>;
  Template?: TemplateResolvers<ContextType>;
  Webhook?: WebhookResolvers<ContextType>;
  WebhookDelivery?: WebhookDeliveryResolvers<ContextType>;
  WebhookStats?: WebhookStatsResolvers<ContextType>;
};